+++ test/core-options-verify/verifySkipIfOlder.test.js
@@ -50,6 +50,24 @@ test('verifySkipIfOlder/undefined', () => {
   expect(warnOutput).toContain(msg)
 })
 
+test.each([null, ''])('verifySkipIfOlder/blank', (value) => {
+  const workingObject = {
+    userOption: { skipIfOlder: value },
+    validatedOption: {}
+  }
+  const warnOutput = []
+  global.console.warn = jest.fn().mockImplementation((s) => { warnOutput.push(s) })
+
+  expect(cov.verifySkipIfOlder(workingObject))
+    .resolves
+    .toMatchObject({
+      userOption: { skipIfOlder: value },
+      validatedOption: { skipIfOlder: DEFAULT_SKIPIFOLDER }
+    })
+  const msg = util.format('Warning: skipIfOlder undefined, defaulting to %s.', DEFAULT_SKIPIFOLDER)
+  expect(console.warn).toBeCalled()
+  expect(warnOutput).toContain(msg)
+})
 test.each(['ANC', '3453', 'xxx', 567])('verifySkipIfOlder/not-boolaen', (value) => {
   const workingObject = {
     userOption: { skipIfOlder: value },
@@ -72,6 +90,25 @@ test('verifySkipIfOlder/undefined/quiet', () => {
     .resolves
     .toMatchObject({
       userOption: {},
+      validatedOption: {
+        skipIfOlder: DEFAULT_SKIPIFOLDER,
+        quiet: true
+      }
+    })
+  expect(console.warn).not.toBeCalled()
+})
+
+test.each([null, ''])('verifySkipIfOlder/blank/quiet', (value) => {
+  const workingObject = {
+    userOption: { skipIfOlder: value },
+    validatedOption: { quiet: true }
+  }
+  global.console.warn = jest.fn()
+
+  expect(cov.verifySkipIfOlder(workingObject))
+    .resolves
+    .toMatchObject({
+      userOption: { skipIfOlder: value },
       validatedOption: {
         skipIfOlder: DEFAULT_SKIPIFOLDER,
         quiet: true